// src/app/api/models/super-admin.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { Address } from '../models/address';
import { UserDto } from '../models/user-dto';

/**
 * Detalles del SuperAdmin a crear.
 */
export interface SuperAdmin {

  /**
   * Objeto Dirección Completo (Para creación inline)
   */
  address?: Address;

  /**
   * Apellido Materno de la Persona
   */
  apellidoMaterno?: string;

  /**
   * Apellido Paterno de la Persona
   */
  apellidoPaterno: string;

  /**
   * Celular de la Persona
   */
  celular?: string;

  /**
   * Tipo de Documento de Identidad
   */
  documentType?: 'DNI' | 'PASSPORT' | 'FOREIGN_ID' | 'HEALTH_CARD' | 'DRIVER_LICENSE' | 'SOCIAL_SECURITY' | 'OTHER';

  /**
   * Correo electrónico personal
   */
  email?: string;
  estadoCivil?: 'SOLTERO' | 'CASADO' | 'UNION_LIBRE' | 'SEPARADO' | 'DIVORCIADO' | 'VIUDO';

  /**
   * La fecha de nacimiento de la Persona
   */
  fechaNacimiento?: string;
  genero?: 'MASCULINO' | 'FEMENINO' | 'OTRO' | 'DESCONOCIDO';

  /**
   * Indica si el SuperAdmin tiene acceso global
   */
  globalAccess: boolean;

  /**
   * Id de la Persona
   */
  id?: number;

  /**
   * Nivel del SuperAdmin
   */
  level: 'LEVEL_1' | 'LEVEL_2' | 'LEVEL_3';

  /**
   * Identificador Nacional de Salud
   */
  nationalHealthId?: string;

  /**
   * Identificador Nacional
   */
  nationalId?: string;

  /**
   * Nombre de la Persona
   */
  nombre: string;
  personTenantsIds?: Array<number>;

  /**
   * Tipo de Persona
   */
  personType?: 'DOCTOR' | 'NURSE' | 'PATIENT' | 'EMPLOYEE' | 'SUPER_ADMIN' | 'ADMIN' | 'RECEPTIONIST' | 'LAB_TECHNICIAN' | 'PHARMACIST';

  /**
   * Segundo nombre de la Persona
   */
  segundoNombre?: string;

  /**
   * Telefono de la Persona
   */
  telefono: string;

  /**
   * Datos del Usuario asociado
   */
  user?: UserDto;
}